refactor(examples): move UIx examples under examples/uix/ (rf2-kx74)
Pairs the substrate-grouped layout introduced for Reagent in the
previous commit with the UIx adapter's smoke-test trio per Decision 7
of Spec 006 §Substrate-adapter shipping convention.

Renames:
  examples/counter_uix/  ->  examples/uix/counter/
  examples/login_uix/    ->  examples/uix/login/

Spec require paths for the shared spec-helpers module are updated to
the new depth. The orchestrator's htmlSrc paths and the shadow-cljs
source paths are wired up in the path-references sweep that follows.
